there are bugs where ? IDK

+++ src/main.rs
@@ -38,8 +38,8 @@ fn handle_connection(mut stream: TcpStream) {
 fn get_altaz_coord(msg : GotoMsg) -> (f64,f64)
 {
     let eq = EqPoint {
-        dec : std::f64::consts::PI * (msg.declination as f64)/(0x1_0000_0000 as f64),
-        asc : std::f64::consts::PI * (msg.declination as f64)/(0x8000_0000 as f64)
+        dec : std::f64::consts::PI * (msg.declination as f64)/((0x1_0000_0000 as u64) as f64),
+        asc : std::f64::consts::PI * (msg.declination as f64)/((0x8000_0000 as u64) as f64)
     };
     convert::eq_to_altaz(eq, 0.795, 0.103)
 }
